Refactored reducers and added onNodeMoved event handler
